--- jni.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <jni.h>
-
-#define JNI_OK = 1; // ???
-
-JNIEnv* getJniEnv(JavaVM *vm);
-bool Jni_Load(const char * libPath)
